refactor: nav link component, nav bar, nav imports, and swap bootstrap for material icons
