+++ client/app.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import NewNav from './components/NewNav'
+import Nav from './components/Nav'
 import Routes from './routes'
 import AuthProvider from './context/authContext'
 
 const App = () => {
   return (
     <AuthProvider>
-      <NewNav />
+      <Nav />
       <Routes />
     </AuthProvider>
   )
