--- client/components/css/navbar.css
@@ -1,56 +1,0 @@
-.navbarContainer {
-  display: flex;
-  position: flex;
-  margin: 0 4em;
-  padding: 10px;
-  width: calc(100vw - 8em);
-  color: #fff;
-  background-color: rgba(37, 37, 37, 0.8);
-  border-radius: 0 0 1em 1em;
-  z-index: 2;
-}
-.heading {
-  flex-grow: 1;
-  padding-top: 0.25em;
-}
-.headerNav {
-  display: flex;
-}
-
-.navItem {
-  height: 1.5em;
-  margin: 0 1.5em 0 0;
-}
-
-.navItem a {
-  font-size: 1.3em;
-  color: #fff;
-  text-decoration: none;
-  padding: 5px;
-}
-
-.navItem a:hover {
-  color: #202b47;
-  border-radius: 5px;
-  background-color: #fff;
-  box-shadow: #d3d3d3;
-}
-
-.navItemLogin {
-  height: 1.5em;
-  margin: 0 1.5em 0 0;
-}
-
-.navItemLogin a {
-  font-size: 1.3em;
-  color: #fff;
-  text-decoration: none;
-  padding: 5px;
-}
-
-.navItemLogin a:hover {
-  color: #202b47;
-  border-radius: 5px;
-  background-color: black;
-  box-shadow: #d3d3d3;
-}
